booking_service/requirements.txt
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
requests==2.32.3
grpcio==1.68.1
grpcio-tools==1.68.1
protobuf==5.28.3